Fix security issue: buffer overflow in mdns message processing.
Malicious MDNS messages had the potential to overflow the static buf allocated
in parse_question because specail_strcpy didn't know about or respect the
size of the buffer.

We fix this by passing the buffer length to special_strcpy and returning -1
if we overflow the buffer.

Thanks to Joel Sandin for reporting this issue.